@@ -115,8 +115,8 @@
     "@electron-forge/plugin-webpack": "^7.11.1",
     "@electron-forge/publisher-github": "^7.11.1",
     "@marshallofsound/webpack-asset-relocator-loader": "^0.5.0",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.3.28",
+    "@types/react-dom": "^18.3.7",
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^4.0.1",
@@ -142,8 +142,8 @@
     "electron-squirrel-startup": "^1.0.0",
     "framer-motion": "^4",
     "immer": "^9.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.2.0",
     "react-redux": "^7.2.3",
     "react-router-dom": "^5.2.0",
