@@ -1,11 +1,18 @@
 import * as React from "react";
-import * as ReactDOM from "react-dom";
-
+import { createRoot } from "react-dom/client";
 import "./App.css";
 import App from "./App";
 
 function render() {
-  ReactDOM.render(<App />, document.getElementById("root"));
+  const container = document.getElementById("root");
+  const root = createRoot(container!);
+  root.render(<App />);
 }
 
 render();
+
+if (module.hot) {
+  module.hot.accept("./App", () => {
+    render();
+  });
+}
